Convert metadata to a list

Takes an object of class xml_document or tibble, and converts it to a list. When converting from an xml_document, this is simply a wrapper for xml2::as_list()

Usage

as_eml_list(x, ...)

# S3 method for class 'tbl_lp'
as_eml_list(x, ...)

# S3 method for class 'tbl_df'
as_eml_list(x, ...)

# S3 method for class 'list'
as_eml_list(x, ...)

# S3 method for class 'xml_document'
as_eml_list(x, ...)

Arguments

x

Object to be converted

...

Other arguments, currently ignored

Value

A list, where both the nested structure of the XML/md and the attributes of XML nodes, are preserved.

Examples

if (FALSE) { # \dontrun{
use_metadata("example_metadata.Rmd")
df <- read_md("example_metadata.Rmd")
df_list <- as_eml_list(df)
} # }